team-stack/web-server/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
PyJWT==2.9.0
